RD-4423: Refactor
